Fixing ino_t in Linux

+++ include/kernel/linux/dirent.h
@@ -10,7 +10,7 @@
 /* types */
 # ifndef ino_t
 #  define ino_t ino_t
-typedef unsigned long long ino_t;
+typedef unsigned long ino_t;
 # endif
 # ifndef off_t
 #  define off_t off_t
